Work on converting package to ROS2
Distribution Statement A; OPSEC #2893
