+++ gps_msgs/CHANGELOG.rst
