+++ gps_msgs/msg/GPSStatus.msg
@@ -1,4 +1,4 @@
-Header header
+std_msgs/Header header
 
 # Satellites used in solution
 uint16 satellites_used # Number of satellites
