+++ gps_msgs/package.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0"?>
+<package format="3">
+  <name>gps_msgs</name>
+  <version>0.3.0</version>
+  <description>GPS messages for use in GPS drivers</description>
+
+  <maintainer email="preed@swri.org">P. J. Reed</maintainer>
+  <license>BSD</license>
+  <url type="website">http://ros.org/wiki/gps_common</url>
+  <url type="repository">https://github.com/swri-robotics/gms_umd</url>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>rosidl_default_generators</buildtool_depend>
+
+  <depend>std_msgs</depend>
+
+  <exec_depend>rosidl_default_runtime</exec_depend>
+
+  <member_of_group>rosidl_interface_packages</member_of_group>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
