+++ gps_tools/package.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0"?>
+<package format="3">
+  <name>gps_tools</name>
+  <version>0.3.0</version>
+  <description>GPS routines for use in GPS drivers</description>
+
+  <maintainer email="preed@swri.org">P. J. Reed</maintainer>
+  <license>BSD</license>
+  <url type="website">http://ros.org/wiki/gps_common</url>
+  <url type="repository">https://github.com/swri-robotics/gms_umd</url>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>ament_cmake_python</buildtool_depend>
+
+  <depend>gps_msgs</depend>
+  <depend>nav_msgs</depend>
+  <depend>rclcpp</depend>
+  <depend>rclcpp_components</depend>
+  <depend>rclpy</depend>
+  <depend>sensor_msgs</depend>
+  <depend>std_msgs</depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
