+++ gps_umd/package.xml
@@ -14,7 +14,8 @@
   <buildtool_depend>catkin</buildtool_depend>
 
   <run_depend>gpsd_client</run_depend>
-  <run_depend>gps_common</run_depend>
+  <run_depend>gps_msgs</run_depend>
+  <run_depend>gps_tools</run_depend>
 
   <export>
     <metapackage/>
