--- gpsd_client/.gitignore
@@ -1,4 +1,0 @@
-/.project
-/.cproject
-/.pydevproject
-/.settings
